rename .js -> .ts, no conversion yet
