+++ src/background-camera.ts
