+++ src/background.ts
